Ver 1.5.1

+++ verify/AOJ-DSL-2-G.test.cpp
@@ -5,7 +5,7 @@
 int main(){
     int n, q;
     cin >> n >> q;
-    BinaryIndexedTreeDouble<int> bit(n);
+    BinaryIndexedTreeDouble<long long> bit(n);
     while(q--){
         int query;
         cin >> query;
